pyproject.toml
[project]
name = "showcase"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio>=6.1.0",
    "langchain>=1.1.3",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.1",
    "python-dotenv>=1.2.1",
]
